requirements.txt
numpy
Pillow
opencv-python
spconv-cu120 
kornia==0.6.12
mmcv==1.5.0
wandb 
albumentations 
imgaug 
scikit-image 
scipy
yacs
fvcore
timm
huggingface_hub